reorg of constants
