+++ MIDAS/src/gnc/constants.h
@@ -1,0 +1,10 @@
+// constants
+const float pi = 3.14159268;
+const float a = 343.0;                // (m/s) speed of sound
+const float rho = 1.225;              // average air density
+const float r = 0.03935;              // (m)
+const float height_full = 4.457;      // (m) height of rocket Full Stage
+const float height_sustainer = 2.029; // (m) height of rocket Sustainer
+const float mass_full = 33.84;        // (kg) Sustainer + Booster
+const float mass_sustainer = 10.93;   // (kg) Sustainer
+const float grav_ms = 9.81;           // (m/s) accel due to gravity
