+++ MIDAS/src/gnc/ekf.cpp
@@ -1,129 +1,13 @@
 #include "ekf.h"
 #include "finite-state-machines/fsm_states.h"
 
+extern const std::map<float, float> O5500X_data;
+extern const std::map<float, float> M685W_data;
+
 EKF::EKF() : KalmanFilter()
 {
     state = KalmanData();
 }
-
-// constants
-const float pi = 3.14159268;
-const float a = 343.0;                // (m/s) speed of sound
-const float rho = 1.225;              // average air density
-const float r = 0.03935;              // (m)
-const float height_full = 4.457;      // (m) height of rocket Full Stage
-const float height_sustainer = 2.029; // (m) height of rocket Sustainer
-const float mass_full = 33.84;        // (kg) Sustainer + Booster
-const float mass_sustainer = 10.93;   // (kg) Sustainer
-
-typedef struct
-{
-    float mach;
-    float alpha;
-    float CA_power_on;
-    float CN;
-    float CP;
-} AeroCoeff;
-
-// stores the aerodynamic coefficients for the corresponding Mach number
-const AeroCoeff aero_data[] = {
-    {0.04, 0, 1.000001789, 25.80486518, 123.856999},
-    {0.08, 0, 0.899149955, 25.80486518, 123.856999},
-    {0.12, 0, 0.848262793, 25.80486518, 123.856999},
-    {0.16, 0, 0.815490497, 25.80486518, 123.856999},
-    {0.2, 0, 0.791977907, 25.80486518, 123.856999},
-    {0.24, 0, 0.77407759, 25.80486518, 123.856999},
-    {0.28, 0, 0.763046286, 25.80486518, 123.856999},
-    {0.32, 0, 0.758319846, 25.80486518, 123.856999},
-    {0.36, 0, 0.760511343, 25.80486518, 123.856999},
-    {0.4, 0, 0.763737136, 25.80486518, 123.856999},
-    {0.44, 0, 0.767325178, 25.80486518, 123.856999},
-    {0.48, 0, 0.771334851, 25.80486518, 123.856999},
-    {0.52, 0, 0.775843406, 25.80486518, 123.856999},
-    {0.56, 0, 0.780953377, 25.80486518, 123.856999},
-    {0.6, 0, 0.785771581, 25.80486518, 123.856999},
-    {0.64, 0, 0.793730593, 25.80486518, 123.856999},
-    {0.68, 0, 0.80285965, 25.80486518, 123.856999},
-    {0.72, 0, 0.807910063, 25.80486518, 123.856999},
-    {0.76, 0, 0.807403195, 25.80486518, 123.856999},
-    {0.8, 0, 0.806889479, 25.80486518, 123.856999},
-    {0.84, 0, 0.832707826, 25.80486518, 123.856999},
-    {0.88, 0, 0.858519521, 25.80486518, 123.856999},
-    {0.92, 0, 0.895125486, 25.492166, 124.3408619},
-    {0.96, 0, 0.923744595, 24.86676763, 125.3085876},
-    {1, 0, 0.941214699, 24.24136926, 126.2763132},
-};
-
-// Number of entries
-#define AERO_DATA_SIZE (sizeof(aero_data) / sizeof(aero_data[0]))
-
-// Moonburner motor thrust curve (Sustainer)
-const std::map<float, float> M685W_data = {
-    {0.083, 1333.469},
-    {0.13, 1368.376},
-    {0.249, 1361.395},
-    {0.308, 1380.012},
-    {0.403, 1359.068},
-    {0.675, 1184.53},
-    {1.018, 1072.826},
-    {1.456, 996.029},
-    {1.977, 958.794},
-    {2.995, 914.578},
-    {3.99, 856.399},
-    {4.985, 781.929},
-    {5.494, 730.732},
-    {5.991, 679.534},
-    {7.258, 542.231},
-    {7.862, 463.107},
-    {8.015, 456.125},
-    {8.998, 330.458},
-    {9.993, 207.118},
-    {10.514, 137.303},
-    {11.496, 34.908},
-    {11.994, 0.0}};
-
-// O5500X motor thrust curve (Booster)
-std::map<float, float> O5500X_data = {
-    {0.009, 20.408},
-    {0.044, 7112.245},
-    {0.063, 6734.694},
-    {0.078, 6897.959},
-    {0.094, 6612.245},
-    {0.109, 6765.306},
-    {0.125, 6540.816},
-    {0.147, 6581.633},
-    {0.194, 6520.408},
-    {0.35, 6795.918},
-    {0.428, 7091.837},
-    {0.563, 7285.714},
-    {0.694, 7408.163},
-    {0.988, 7581.633},
-    {1.266, 7622.449},
-    {1.491, 7724.49},
-    {1.581, 7653.061},
-    {1.641, 7540.816},
-    {1.684, 7500.0},
-    {1.716, 7336.735},
-    {1.784, 7224.49},
-    {1.938, 6785.714},
-    {2.138, 6326.531},
-    {2.491, 5897.959},
-    {2.6, 5704.082},
-    {2.919, 3540.816},
-    {3.022, 3408.163},
-    {3.138, 2887.755},
-    {3.3, 2234.694},
-    {3.388, 1673.469},
-    {3.441, 1489.796},
-    {3.544, 1418.367},
-    {3.609, 1295.918},
-    {3.688, 816.327},
-    {3.778, 653.061},
-    {3.819, 581.633},
-    {3.853, 489.796},
-    {3.897, 285.714},
-    {3.981, 20.408},
-    {3.997, 0.0}};
 
 /**
  * @brief Sets altitude by averaging 30 barometer measurements taken 100 ms
@@ -180,6 +64,7 @@ void EKF::initialize(RocketSystems *args)
 
     F_mat.setZero(); // Initialize with zeros
 
+    // Initialize Q from filterpy
     Q(0, 0) = pow(s_dt, 5) / 20;
     Q(0, 1) = pow(s_dt, 4) / 8;
     Q(0, 2) = pow(s_dt, 3) / 6;
@@ -244,37 +129,43 @@ void EKF::priori(float dt, Orientation &orientation, FSMState fsm)
     // Eigen::Matrix<float, 3, 1> gravity = Eigen::Matrix<float, 3,1>::Zero();
     if ((fsm > FSMState::STATE_IDLE) && (fsm < FSMState::STATE_LANDED))
     {
-        gravity(0, 0) = -9.81;
+        gravity(0, 0) = -grav_ms;
     }
     else
     {
         gravity(0, 0) = 0;
     }
-    float m = mass_sustainer;
-    float h = height_sustainer;
+
+    // mass and height initialization
+    float curr_mass_kg = mass_sustainer;
+    float curr_height_m = height_sustainer;
 
     if (fsm < FSMState::STATE_BURNOUT)
     {
-        m = mass_full;
-        h = height_full;
+        curr_mass_kg = mass_full;
+        curr_height_m = height_full;
     }
 
     float w_x = omega.vx;
     float w_y = omega.vy;
     float w_z = omega.vz;
 
-    float vel_mag_squared = x_k(1, 0) * x_k(1, 0) + x_k(4, 0) * x_k(4, 0) + x_k(7, 0) * x_k(7, 0);
+    // finding the Mach number
+    float vel_mag_squared_ms = x_k(1, 0) * x_k(1, 0) + x_k(4, 0) * x_k(4, 0) + x_k(7, 0) * x_k(7, 0);
+    float vel_magnitude_ms = pow(vel_mag_squared_ms, 0.5);
+    float mach = vel_magnitude_ms / a;
 
-    float velocity_magnitude = pow(vel_mag_squared, 0.5);
-    float mach = velocity_magnitude / a;
+    // finding the closest aerodynamic coeff. C_a
     int index = std::round(mach / 0.04);
     index = std::clamp(index, 0, (int)AERO_DATA_SIZE - 1);
     Ca = aero_data[index].CA_power_on;
 
-    float Fax = -0.5 * rho * (vel_mag_squared) * float(Ca) * (pi * r * r);
-    float Fay = 0;
-    float Faz = 0;
+    // aerodynamic force
+    float Fax = -0.5 * rho * (vel_mag_squared_ms) * float(Ca) * (pi * r * r);
+    float Fay = 0; // assuming no aerodynamic effects
+    float Faz = 0; // assuming no aerodynamic effects
 
+    // force due to gravity
     Eigen::Matrix<float, 3, 1> Fg_body;
     EKF::GlobalToBody(angles, Fg_body);
 
@@ -282,25 +173,28 @@ void EKF::priori(float dt, Orientation &orientation, FSMState fsm)
     float Fgy = gravity(1, 0);
     float Fgz = gravity(2, 0);
 
-    Eigen::Matrix<float, 3, 1> T;
+    // thurst force
+    Eigen::Matrix<float, 3, 1> thrust_vec;
 
-    EKF::getThrust(stage_timestamp, angles, fsm, T);
+    EKF::getThrust(stage_timestamp, angles, fsm, thrust_vec);
 
-    float Ftx = T(0, 0);
-    float Fty = T(1, 0);
-    float Ftz = T(2, 0);
+    float Ftx = thrust_vec(0, 0);
+    float Fty = thrust_vec(1, 0);
+    float Ftz = thrust_vec(2, 0);
 
     xdot << x_k(1, 0),
-        ((Fax + Ftx + Fgx) / m - (w_y * x_k(7, 0) - w_z * x_k(4, 0)) + x_k(2, 0)) * 0.5,
+        ((Fax + Ftx + Fgx) / curr_mass_kg - (w_y * x_k(7, 0) - w_z * x_k(4, 0)) + x_k(2, 0)) * 0.5,
         0.0,
 
         x_k(4, 0),
-        ((Fay + Fty + Fgy) / m - (w_z * x_k(1, 0) - w_x * x_k(7, 0)) + x_k(5, 0)) * 0.5,
+        ((Fay + Fty + Fgy) / curr_mass_kg - (w_z * x_k(1, 0) - w_x * x_k(7, 0)) + x_k(5, 0)) * 0.5,
         0.0,
 
         x_k(7, 0),
-        ((Faz + Ftz + Fgz) / m - (w_x * x_k(4, 0) - w_y * x_k(1, 0)) + x_k(8, 0)) * 0.5,
+        ((Faz + Ftz + Fgz) / curr_mass_kg - (w_x * x_k(4, 0) - w_y * x_k(1, 0)) + x_k(8, 0)) * 0.5,
         0.0;
+        
+    // priori step
     x_priori = (xdot * dt) + x_k;
     setF(dt, w_x, w_y, w_z);
     P_priori = (F_mat * P_k * F_mat.transpose()) + Q;
@@ -336,7 +230,7 @@ void EKF::getThrust(float timestamp, euler_t angles, FSMState FSM_state, Eigen::
         if (FSM_state < FSMState::STATE_BURNOUT)
         {
             // first stage
-            if (timestamp >= 0.009)
+            if (timestamp >= 0.009) // TODO: index the first value from the correct motor
             {
                 auto it = O5500X_data.lower_bound(timestamp);
                 if (it != O5500X_data.end())
@@ -352,7 +246,7 @@ void EKF::getThrust(float timestamp, euler_t angles, FSMState FSM_state, Eigen::
         }
         else
         {
-            if (timestamp >= 0.083)
+            if (timestamp >= 0.083) // TODO: index the first value from the correct motor
             {
                 // second stage
                 auto it = M685W_data.lower_bound(timestamp);
@@ -408,6 +302,7 @@ void EKF::update(Barometer barometer, Acceleration acceleration, Orientation ori
     // Sensor Measurements
     Eigen::Matrix<float, 3, 1> accel = Eigen::Matrix<float, 3, 1>(Eigen::Matrix<float, 3, 1>::Zero());
 
+    // TODO: document this right
     (accel)(0, 0) = acceleration.az - 0.045;
     (accel)(1, 0) = acceleration.ay - 0.065;
     (accel)(2, 0) = -acceleration.ax - 0.06;
@@ -417,16 +312,18 @@ void EKF::update(Barometer barometer, Acceleration acceleration, Orientation ori
 
     Eigen::Matrix<float, 3, 1> acc;
     EKF::BodyToGlobal(angles, accel, acc);
+
     float g;
     if ((FSM_state > FSMState::STATE_IDLE) && (FSM_state < FSMState::STATE_LANDED))
     {
-        g = -9.81;
+        g = -grav_ms;
     }
     else
     {
         g = 0;
     }
 
+    // TODO: magic numbers
     y_k(1, 0) = ((acc)(0)) * 9.81 + g;
     y_k(2, 0) = ((acc)(1)) * 9.81;
     y_k(3, 0) = ((acc)(2)) * 9.81;
